feat: remove old routes, views, controller and policy
