+++ app/policies/settings/personal/profile_policy.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class ATS::ProfilePolicy < ApplicationPolicy
+class Settings::Personal::ProfilePolicy < ApplicationPolicy
   # Not all users will want to link their email accounts,
   # so for now we have decided to hide this functionality.
   def link_gmail?
