--- app/views/ats/settings/_account_info.html.slim
@@ -1,17 +1,0 @@
-/# locals: (account:)
-
-= form_with model: account,
-            url: update_account_ats_settings_path,
-            id: :settings_form,
-            method: :patch,
-            class: "vstack gap-3" do |form|
-
-  = render LabeledComponent.new(left_layout_class: "col-12", right_layout_class: "col-12") do |c|
-    - c.with_label t("user_accounts.settings.full_name_input_label"), for_field: :full_name
-    = render TextInputComponent.new(form,
-                                    method: :name,
-                                    size: :small,
-                                    type: :text,
-                                    required: true,
-                                    value: account.name,
-                                    aria: { label: t("user_accounts.settings.full_name_input_label") })
