--- app/views/ats/settings/show.html.slim
@@ -1,32 +1,0 @@
-- ats_title(t("user_accounts.settings.show.title"))
-.authentication-container.mx-auto.card
-  .card-body.p-4.vstack.gap-3
-    h1 = t("user_accounts.settings.show.title")
-    / Inner card block
-    .vstack.gap-3
-      = render "account_avatar", account: current_account
-      = render "account_info", account: current_account
-      / Email input with a link button
-      = render LabeledComponent.new(left_layout_class: "col-12", right_layout_class: "col-12") do |c|
-        - c.with_label t("user_accounts.settings.show.email_input_label"), for_field: :email
-        .hstack.gap-2
-          = render TextInputComponent.new(:email,
-                                          type: :email,
-                                          readonly: true,
-                                          value: current_account.email,
-                                          aria: { label: t("user_accounts.settings.show.email_input_label") })
-          - if allowed_to?(:link_gmail?, with: ATS::ProfilePolicy)
-            = render ButtonLinkComponent.new(@link_gmail_uri,
-                                             variant: :secondary,
-                                             disabled: current_member.email_service_linked?)
-              - if current_member.email_service_linked?
-                = t("user_accounts.settings.show.email_linked")
-              - else
-                = t("user_accounts.settings.show.email_not_linked")
-      / Action buttons
-      .vstack.gap-2.mt-2
-        .hstack.gap-2
-          = render ButtonComponent.new(class: "col", form: :settings_form)
-            = t("core.save_button")
-        span
-          = link_to t("user_accounts.settings.show.change_password_link_text"), rodauth.change_password_path
