+++ app/views/layouts/ats/_navbar.html.slim
@@ -1,6 +1,6 @@
 - account_icon_dropdown_elements = capture do
     li.nav-item class=("active" if request.fullpath.eql?("/ats/settings"))
-      = link_to t("core.settings"), ats_settings_path, class: "nav-link"
+      = link_to t("core.settings"), settings_personal_profile_path, class: "nav-link"
     li.nav-item
       = button_to t("rodauth.logout_button"), rodauth.logout_path, method: :post, class: "nav-link"
