+++ app/views/rodauth/change_password.html.slim
@@ -38,7 +38,7 @@
         .hstack.gap-2
           = render ButtonComponent.new(class: "col")
             = rodauth.change_password_button
-          = render ButtonLinkComponent.new(ats_settings_path, class: "col", variant: :secondary)
+          = render ButtonLinkComponent.new(settings_personal_profile_path, class: "col", variant: :secondary)
             = t("rodauth.back_button")
         span
           = link_to t("rodauth.reset_password_request_link_text"), rodauth.reset_password_request_path
