+++ app/views/settings/personal/profiles/_account_avatar.html.slim
@@ -8,9 +8,9 @@
       = render IconComponent.new(:user)
   .hstack.gap-2
     = form_with model: account,
-                url: update_avatar_ats_settings_path,
+                url: update_avatar_settings_personal_profile_path,
                 method: :patch, class: "hstack gap-2 turbo-instant-submit" do |form|
       = render "shared/profile/image_upload_button", file_field: :avatar, form:
 
     - if account.avatar.present?
-      = render "shared/profile/image_delete_button", url: remove_avatar_ats_settings_path
+      = render "shared/profile/image_delete_button", url: remove_avatar_settings_personal_profile_path
