+++ app/views/settings/personal/profiles/_account_info.html.slim
@@ -1,7 +1,7 @@
 /# locals: (account:)
 
 = form_with model: account,
-            url: update_account_ats_settings_path,
+            url: update_account_settings_personal_profile_path,
             id: :settings_form,
             method: :patch,
             class: "vstack" do |form|
